feat: Update church dates and descriptions
This commit updates the dates and descriptions of the church events on the" Cronograma" page. The previous dates were incorrect, and the descriptions were not clear. Now, the dates are accurate, and the descriptions provide more information about each culto event. This improves the accuracy and clarity of the information displayed on the page.